refactor: GNB 알람 리팩토링
