--- src/components/GNB/UserGNB/mobile/MobileTILModal/style.ts
@@ -1,36 +1,0 @@
-import { css } from '@emotion/react';
-import styled from '@emotion/styled';
-
-export const ModalContainerStyles = css`
-  width: 90vw;
-`;
-
-export const CloseButtonStyles = () => css`
-  top: 20px;
-  right: 20px;
-`;
-
-export const ModalContentStyles = () => css`
-  padding: 1.5rem 2rem;
-  font-size: 1.375rem;
-  font-weight: 700;
-  width: 100%;
-`;
-
-export const ModalTitle = styled.h3`
-  font-size: 1.5rem;
-  font-weight: 700;
-`;
-
-export const TabStyles = css`
-  margin-top: 1.25rem;
-  height: 2.5rem;
-`;
-
-export const TabMenuStyles = css`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 2rem;
-  font-size: 1.125rem;
-`;
